+++ lib/features/workout/screens/workouts/_base_workout_screen.dart
@@ -3,21 +3,19 @@ import "dart:async";
 import "package:flutter/foundation.dart";
 import "package:flutter/material.dart";
 import "package:logging/logging.dart";
-import "package:lucide_icons_flutter/lucide_icons.dart";
 import "package:provider/provider.dart";
 import "package:pull_up_club/common/providers/workout_history_provider.dart";
 import "package:pull_up_club/common/themes/app_colors.dart";
 import "package:pull_up_club/common/themes/app_spacing.dart";
 import "package:pull_up_club/common/themes/app_typography.dart";
 import "package:pull_up_club/common/utils/utils.dart";
-import "package:pull_up_club/common/widgets/core/gradient_button.dart";
 import "package:pull_up_club/common/widgets/core/screen_scaffold.dart";
-import "package:pull_up_club/common/widgets/shared/home_button.dart";
 import "package:pull_up_club/common/widgets/shared/set_cards.dart";
 import "package:pull_up_club/domain/models.dart";
 import "package:pull_up_club/features/workout/providers/workout_provider.dart";
 import "package:pull_up_club/features/workout/screens/rest_screen.dart";
 import "package:pull_up_club/features/workout/screens/success_screen.dart";
+import "package:pull_up_club/features/workout/widgets/destructive_workout_buttons.dart";
 
 abstract class BaseWorkoutScreen extends StatefulWidget {
   const BaseWorkoutScreen({super.key});
@@ -168,24 +166,7 @@ abstract class BaseWorkoutState<T extends BaseWorkoutScreen> extends State<T> {
             numExpectedCards: workout.maxGroups,
             highlightedIndex: getNumCompletedGroups(),
           ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: [
-              SizedBox(
-                width: Screen.width(context) * 0.4,
-                child: GradientButton(
-                  onPressed: workout.sets.isEmpty ? null : undoLastSet,
-                  text: "Back",
-                  icon: LucideIcons.arrowLeft,
-                  gradient: AppGradients.light,
-                ),
-              ),
-              SizedBox(
-                width: Screen.width(context) * 0.4,
-                child: const HomeButton(text: "Abort", icon: LucideIcons.x),
-              ),
-            ],
-          ),
+          DestructiveWorkoutButtons(workout: workout, onUndoLastSet: undoLastSet),
         ],
       ),
     );